requirements/base.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

django==1.8.13
django-sekizai==0.9.0
psycopg2>=2.6.1,<2.7
django-model-utils==2.3.1
django-fontawesome==0.2.6
django-bootstrap3==7.0.0
django-authtools==1.5.0

